LayoutTests/fullscreen/full-screen-video-offset.html

<body onload="init()">
    <div>This layout test checks that the offset size and position of a video element does not change when the it enters full-screen mode.  Press <button onclick="video.webkitRequestFullScreen()">go full-screen</a> to begin.</div>
    <video id="video" controls></video>
    <script src="full-screen-test.js"></script>
    <script>
        var video = document.getElementById('video');
        function init() {

            // Bail out early if the full screen API is not enabled or is missing:
            if (Element.prototype.webkitRequestFullScreen == undefined) {
                logResult(false, "Element.prototype.webkitRequestFullScreen == undefined");
                endTest();
            } else {
                var videoState = {'offsetLeft': video.offsetLeft, 'offsetTop': video.offsetTop, 'offsetWidth': video.offsetWidth, 'offsetHeight': video.offsetHeight };

                var fullscreenChanged = function(event)
                {
                    testExpected("document.webkitCurrentFullScreenElement", video);
                    testExpected("video.offsetLeft", videoState.offsetLeft);
                    testExpected("video.offsetTop", videoState.offsetTop);
                    testExpected("video.offsetWidth", videoState.offsetWidth);
                    testExpected("video.offsetHeight", videoState.offsetHeight);
                    endTest();
                };
                waitForEvent(document, 'webkitfullscreenchange', fullscreenChanged);    
                runWithKeyDown(function(){video.webkitRequestFullScreen()});
            }
        }
    </script>
</body>
